polished up forest section
changed some pathing, added stone material to landscape brush and painted on edges, added dead trees to the edges, some stone, converted meshes to nanite and is very easy to convert back. I'm not to sure about the edges of the forest level still but we can discuss that Monday.